requirements.txt
httpx==0.27.0
geopy==2.4.1
Django==3.2.23
python-dotenv==1.0.0
polyline==2.0.2
folium==0.20.0
legacy-cgi==2.6.3; python_version >= '3.13'
pre-commit==4.2.0
